Add Employees module new conf settings
It was added a left menu to module just as example, the index file was added as a copy of topmenu file because it was changed the target of the top menu button.

I notice that module purpose is only for vendors, so it's necessary to create a new specific function module just to avoid confusions in the future.