Create general sales report table
It was included a table with the amount and debit of each vendor. The query here doesn't consider any date range restriccion, so it will be included when the period input is ready.

When the query is completed, it must exports the same data on table in an Excel file.